Added note for gateway name and port
As added in the part of creatig the transparant gateway. It needs to be clear that the gateway name needs to be the same all around and that it needs to be resolvable to an IP Address. And also it needs to be clear that when there is a firewall between the downstream device and the transparant gateway, the respective port needs to be open.