M33568: Changing string to match UI
Hello, @bandersmsft,
Translator has reported possible source content issue. 
Description: "Weekly costs for the last seven days" should be not correct because it is indicated in the UI "Weekly export of costs for the last 7 days" in the screenshot on the live web page (https://docs.microsoft.com/en-us/azure/cost-management/tutorial-export-acm-data). 


Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.