Update webjobs-sdk-get-started.md
The guide tells you to create a .NET Framework console application, and this project template doesn't create an appsettings.json file, so you have to add it manually. However, you'll run into an issue when you get to the section "Test locally", which is an InvalidOperationException telling you that the storage account "Storage" has not been configured. It turns out that the appsettings.json file needs to be copied to the output directory (Which of course makes sense), but the guide doesn't say anything about this step.

So this change is simply adding this step to help others avoid spending time on Googling and debugging.