Reflect the latest both reboot behavior for this documentation.
Change the following sentence
from
"If you reboot both the **Master** and **Slave** nodes, all data in the cache (or in that shard if you are using a premium cache with clustering enabled) is lost."
to
"If you reboot both the **Master** and **Slave** nodes, all data in the cache (or in that shard if you are using a premium cache with clustering enabled) may be lost but this is not guaranteed either. "

I checked the Redis Cache behavior that all data in the cache sometimes lose and sometimes keep if we reboot both node.