Fixed typo in config instructions
Azure Digital Twins' Tutorial on TSI Integration has a typo instructing users to fill out the `actions\createEndpoints.yaml` config that will result in both a 400 error for half and a failed endpoint for the other half. The path variable should be the name of the eventHub, not the name of the namespace the hub lives within.