M35676: Localizability: Azure Files and Folders
Hello @saurabhsensharma,
Translator has reported possible source content issue.
Description: Is "Azure Files and Folders" a service name or should "files and folders" be translated? There is an Azure service called "Azure Files", but there is no service called "Azure Folders" or "Azure Files and Folders." If this string does not refer to an Azure service, then the capitalization should be "Azure files and folders."
Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.