Added the classes you can use for the Recieve method
I have added the classes that you can use the Recieve method on.

Reason for this is I could not find out how to `Recieve` from a subscription. Could not find any documentation anywhere. In the end Intellisense helped me out with this, so hopefully this will help someone else.