Specified the needed permissions
It was not specified anywhere in this document or anywhere else the permissions needed for the managed identity to get this to work.  I propose this doc. change because I spent a long time trying to access the key vault with a Managed Identity having only the GET permission. After adding the LIST permission too, it worked.