As per the comment in the article
As per the comment in the article, the below statement sounded kind of complicated with a lot of comma splits. Changed the statement to be more clear.

Note that both Internet-facing and internal Application Gateways route requests to your backend servers using private IP addresses,if your backend pool resource contains a private IP address, VM NIC configuration, or an internally resolvable address, and if your backend pool is a public endpoint, Application Gateway uses its frontend public IP to reach the server. If you haven't provisioned a frontend public IP address, one is assigned for the outbound external connectivity.