Update concept-password-ban-bad-on-premises.md
MS Staff is telling customers, that password hash sync is required for that feature, so it should made clear in this document that this is not the case. Also password policies from Azure AD with the Well-known passwords list, as well as the traditional password policies in the local domain, are likely to be confused. Tried to clear this up.