Clarified language on stats granularity
Perhaps this is a bit of a style thing, but the sentence I modified in this PR was a bit hard to parse. This exact information was what I was in need of learning, and after a bit of reading I realize the answer was in front of me. I propose a bit of a wording change and an additional sentence to make it as clear as possible what's being communicated.