Update backup-azure-arm-restore-vms.md
Changed the line: "After you trigger the restore operation, the backup service creates a job for tracking. Azure Backup displays notifications about the job in the portal. If they aren't visible, click on the **Notifications** symbol and so, click **View all Jobs** to see the Restore Process Status." Including click in All Jobs due a frequent customers confusion on a missing step.