(Azure CXP) Fixed the Remove license Script as on MicrosoftDocs/azure-docs#30064
Fixed the license remval script . The section **Remove direct licenses for users with group licenses** contains a script which was not working properly as reported by the customer via MicrosoftDocs/azure-docs#30064  . 

The script block was missing a pipe sign due to which the output of the previous cmdlet  `$groupId`. `Get-MsolGroupMember` was not present within the powershell pipeline for the next cmdlet `Get-MsolUser` to use. I have fixed the same and will get this updated after review. The following is the script block for your reference. 

`Get-MsolGroupMember -All -GroupObjectId $groupId | Get-MsolUser -ObjectId {$_.ObjectId} | Foreach {
`

Thanks