C86530: Fixing asterisks issues on localized pages
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.  
Description: There are cases of isolated asterisk shown up on live build. As for line 132 consider using a backslash before the asterisk in case it needs to be as is or adding the closing italic tag.
As for line 133 kindly consider removing the extra asterisk so it keep consistency with the line 132

Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.
Many thanks in advance.