Included the range of prefix under constraints
There's a disconnect between the sample shown on the pricing calculator and what you can actually create as an Azure Public IP Prefix.  I attempted to create a /24 and was greeted with an error that the minimum is /28 and the maximum is /31.  This is inside the Az cloud shell within the portal.  

I added the line here so that customers know what they can and can't deploy without trial and error.