(AzureCXP) https://github.com/MicrosoftDocs/azure-docs/issues/30964
Add  the following lines as doc enhancement: 
"It is not possible to send a notification to more than one namespace. Namespaces are a logical container for notification hubs and are not involved with sending notifications. The namespace-level access policies (credentials) can be used for namespace-level operations, for example: listing notification hubs, creating or deleting notification hubs, etc. Only the hub-level access policies would let you send notifications."