Added Azure AD to Role name
When reading this document it is confusing when it starts talking about the "Applicaiton Developer" role, after spending the last section talking about RBAC permissions. There is no indicator that this is an Azure AD role, seperate to RBAC, and I have seen users be confused by this. I have added clarification.