Included links to the CLI and PS cmdlets
Included links to the CLI and PS cmdlets that allow you to programmatically consume the ip addresses in the prefix you've just created.  This helps the end user by not having to search for the appropriate commands.  This also follows the same guidelines as the paragraph above that shows you the cli and ps for programmatically creating a Public IP prefix.