Add limitation to the list
Adding the limitation: Columns with User Defined Data Types are not supported
There is open feedback about this topic:
https://feedback.azure.com/forums/908035-sql-server/suggestions/32900872-allow-columns-with-user-defined-data-types-to-be-i
At this time Microsoft team review this option. In the mean time people ask about this in the MSDN forums and since it was not yet documented they present it as a bug.