Including language around UDR support for V2 SKU
The question around "why doesn't V2 sku support UDR" has been asked more than one time at the partner level, in discussions with Microsoft.  Each time someone references this document, and someone from PM states it's on a backlog.  We can't really depend on that statement here.  The effort is to clarify for folks that UDRs are not supported at General Availability, which hopefully leads one to believe that it is possible the support will show up post-GA, and instead ask the question of when.  Rather than reading the statement as finite, that it is not supported and no one is working on bringing it to v2 SKU.  The note below it states why using UDRs is not a good practice if you want back-end health, logs, and metrics, but people will do what the system allows, at their level of risk acceptance. 

This also makes the statement that if you require UDR you shouldn't move to v2 (with the expectation that support will show up any time soon).  That you should continue with v1.  This would be the correct "practice" for customers until public committments of support are made by Microsoft.