Update virtual-network-network-interface.md
The reason I am adding this change is because this feature appears undocumented at this point. When you change a NIC that is part of an Availability set the portal pops up this notice "If the VM that uses a NIC which is part of an availability set, then all the DNS servers specified for each of the VMs from all NICs that are part of the availability set will be inherited." but this is not mentioned in the documentation. Please confirm and update the doc.