Update the Delegated Login Identity option
This doc categorically says the answer is "SAM Account Name" - but it is not. This is not the correct configuration if you have a multi-domain AD forest. It took a few weeks to work out what of all the settings was not working and we found it was this. The App Proxy SAM Account Name does not include the domain name, and so it will try to login to SharePoint with "username" and not "domain\username" which it needs to in a multi-domain environment.