Remove implication quota increases are possible
Within Notification Hubs we cannot currently grant quota increase on device counts due to the strain it puts on the service and other regional customers. This commit removes from the FAQ the implication that we have the ability to increase quotas for customers.

Our recommendation for customers today is to partition their devices across multiple hubs.