Update in setting the network host mode
We encountered issues when using OPC Twin module with createOptions as defined in your sample. The error message was this through edgeAgent logs: 

System.AggregateException: One or more errors occurred. (Error calling start module opctwin: Could not start module opctwin
        caused by: Could not start module opctwin
        caused by: failed to add interface veth15550b3 to sandbox: error setting interface "veth15550b3" IP to [My IP...]/16: cannot program address [My IP...]/16 in sandbox interface because it conflicts with existing route {Ifindex: 4 Dst: [my IP] Src: [my IP] Gw: <nil> Flags: [] Table: 254}) ---> Microsoft.Azure.Devices.Edge.Agent.Edgelet.EdgeletCommunicationException: Error calling start module opctwin: Could not start module opctwin
=====

Setting the host mode with the new createOptions in this update works on Ubuntu 16 LTS as well as ARM32 Raspbian. I have not tested on other platforms though. Not sure if that would negatively affect those.