Improve language around persistence
The previous version implies that the only place data is persisted is the clouddrive file share, but https://docs.microsoft.com/en-us/azure/cloud-shell/persisting-shell-storage states the situation more correctly and clearly. This edit attempts to remove the incorrect implication without going into depth that is already on the linked-to page.