requesting limitation to be posted near line 60-61
Requesting edit to add information on a limitation for the content pack being used on government tenants around lines 60 and 61. This information was confirmed by my contact with Dhanyah Krishnamoorthy. Feel free to contact me to get more info if required. The best way to contact me is through teams or email via v-thsimp@microsoft.com