Cleaning up auth methods for AAD Database users and Logins
"Azure Active Directory Integrated" was listed twice for aad logins.  "Using an Azure AD principal name and a password" should be changed to "Azure Active Directory Password" , and "Using integrated Windows authentication" changed to "Azure Active Directory Integrated".  And is it really the case that token auth is not supported for AAD Logins?