Limitation for  Approved Client App (Native App)
Here the term Approved Client App is also called as Native App. Please note, a native application is a software program that is developed for use on a specific platform or device. For example: - Kaizala App that comes under the category of Native App, that is designed for IOS and Android Devices.  The configuration of Native App is totally different as compare to the Cloud Apps like Microsoft Teams, Yammer, Skype for Business online. That also means that any Native App (such as Kaizala) does not qualifies for a mechanism to restrict through Conditional Access, unlike the policies applied for a Cloud App.