(AzureCXP) Added note for xml encoding
Added please note that the client secret will need to be XML encoded to properly work in your web.config. 

As it was confusing as to why the secret wasn't properly working when copying and pasting into the web.config. 

I found out this issue when debugging git issue : https://github.com/MicrosoftDocs/azure-docs/issues/35232 
So this might be an issue that is being trickled down that is causing the issue linked.