FIxed self referencing confusion
I have just implemented this product for a client of mine and found the instructions very confusing (and the ones Adaptive provide are also not good - and probably just a copy and paste of a previous version of this page!). To get the "Basic SAML Configuration" URLs you need to do stuff in Adaptive, but that stuff is unknown until you do the settings in AzureAD - so self referencing. By placing placeholder values for all Adaptives required settings you can get what you need for Azure AD, and then you go back and fix Adaptive.