Updated text
Updated text, not sure how we want to include this into the buildingblocks. There seem to be alot of issues with the B2C docs, and this is only adding one element that I know of. It seems like per the doc in the issue provided, there might be more.