Update sql-database-ssms-mfa-authentication.md
Based on a conversation with a colleague, that works as Identity PFE in the SI Domain (mdiiorio), we have had a look at the following documentation: https://docs.microsoft.com/en-gb/azure/security/azure-ad-choose-authn (reference is near the Decision Tree Diagram) and have found that Azure AD does not natively supported sign-in using smartcards or certificates. Can you please check if this capability is supported or if we should modify the text to highlight that is a federated domain?