Please add these two question
Q. If I block user sign-in at O365/AAD Portal for synced users, why it is unblock the sign-in back?.
For synced users, the users are authored/mastered from on-prem. hence if the account is enabled at on-prem it will unblock the sign-in block placed by admin
Q. If I disable the synced Device (ie. HAADJ) at Azure portal , why it is enabled back?
For Synced devices, Devices are authored/mastered from on-prem. hence if the device is enabled at on-prem it will enable the device at AAD portal which was previously disabled by admin.
To disable synced device, Please use on-prem AD side to disable the computer account.