Updated accurate data on Azure Web Storage Explorer
As the mantainer of the project I just wanted to provide accurate data about it.
It is a web app (not Windows like it said) although there's also a Docker image with it so eventually, you could have it running on your box (any box). Also, Files are also available, so I checked that feature in the table.