Add Hillstone VPN device information
Hi,

This is Yi from Hillstone Networks. I see my pull request was closed. If you need to start the validation process, please let us know. We also have office and colleagues in the Bay Area. You can contact yizhao@hillstonenet.com and yzhou@Hillstonenet.com. Btw. As I don't see the Email of author on the profile page, I can only leave message on this thread. Not sure if it can reach the author in my previous request as it's closed. Could I have the contact info of the author for the following validation process if possible?

Thanks,
Yi