Update dns-reverse-dns-for-azure-services.md
I changed "add' to "update" because the commands listed will not work because the powershell result for Get-AzPublicIpAddress will not have a field .ReverseFqdn if there is no value already set for the .ReverseFqdn. Then this command fails. So changing add to update seems logical here. Please check on your end, or we can add a note here.