Updated with IRS Control numbers
@DCtheGeek - I also deleted link to the deployment steps under section "Next Steps" since this will be only the Initiative for now. 
Please review and provide me any feedback. Then once the initiative is live in the portal then we will press publish on this content. There won't be a blog since this is only an initiative. I have a completed approved PR but I'm doing the final merge step. If you need the link let me know.