Linking to a complete sample for Web App calling Web API
The existing link takes you to a page with a lot of good reference material, but it's a horrible tutorial (e.g. there is no start or end state and there is very little how-to involved). Instead, I'm proposing to change the link to the very well documented, complete, and working tutorial for Microsoft Graph ("Build ASP.NET MVC apps with Microsoft Graph"). This also helps users in the community start working with the Graph as their sample API.