Modified bullets under Cluster name based on CSS case
One of the customers had the following question - What is the scope within which the HDI cluster name needs to be unique? 
The answer provided by CSS : The cluster naming scope is for all Azure, across all subscriptions. So you cluster name must be unique worldwide. - so updated this doc to reflect this so that we can deflect the same questions in future.