Update starter.md changed "limit" to "top"
This query will use `limit` to only retrieve five matching records that are ordered by name. The type
of the Azure resource is `Microsoft.Compute/virtualMachines`. `project` tells Azure Resource Graph
which properties to include.

Limit just limits the output, top allows for sorting as provided in the example as well. Changed "limit" to "top" in the introductory paragraph as this is used in the examples as well.