Fixed the indentation for the F#
Fixed the indentation for the Paket.Package that was throwing error:  ```This value is not a function and cannot be applied. Did you forget to terminate a declaration?``` when copied and pasted directly into the F# notebook.
Added 2 extra steps to make it clearer on how to get the libraries functioning in the notebook.