Update network-watcher-packet-capture-manage-cli.md
I added a command for showing the Network Watcher extension for a Linux VM, as the difference in naming (e.g. AzureNetworkWatcherExtension for Linux) causes confusion when referencing the name of the extension on Windows.  Additionally, I may be mistaken, but I think the existing Step 1 under "Start a packet capture" is unnecessary.  "az network watcher show" is invalid.  Also, I changed "Start a packet capture" step 2 to retrieve the storage account with "az storage account list" instead of "azure storage account list".  Finally, I changed references to 'cmdlet' to 'command' to make it more consistent for the CLI portion of the documentation.