Examples 2 and 3 calculation seem incorrect
For the example 2 and 3 we're not including the minimum number of instances in auto-scale mode, in the calculations.
As we say later on the document:
"Minimum capacity ensures that Application Gateway and WAF v2 don't fall below the minimum instance count specified, even in the absence of traffic. You'll be billed for this minimum capacity even in the absence of any traffic"

So, even if there was only one instance on the VMSS, 10 compute units need to be included in the max() calculations, which for these examples will superseed the other values under the Max formula.

Likley we need to add examples with more consumption to clearly explain the differences, but we shouldn't exclude the minimum consumption as it's driving confusion among customers and leading to case opening.