Restoring file deleted by my predecessor
A few months ago we retired a few features in Security Center. This retirement was completed a few weeks ago and my predecessor removed all relevant articles from the documentation (Commit: 9dc6fa56c35bb86c948e36bf27871e1875511c6e and 15b99cebd8afdeb459a93d598cfcecce0a96b142).
She accidentally removed this article which is still a feature in ASC (only a small part of it was retired, and the docs state that).
At the request of the product team, I'm restoring this file.