Some sentences about ER and VPN coexisting are confusing.
I know this document is about how to add a vpn connection to existing VPN Gateway. We can use this document on ER and VPN coexisting environment. But there are some sentences which tell that we can not use this document on ExpressRoute and Site-to-Site coexisting environment. I think you should delete those sentences or revise them. Could you check it ?