Fixed pre-requisites
It is not required to be an azure active directory global administrator to get Azure Security Centre log information. Also it is required to be an Owner of the subscription for this connector to work and it is not accurate to state that this connector needs Azure AD Global administrator permissions. I have updated the document with correct permissions needed for this connector to work.