Databricks Runtime Version 5.2 no longer available
Runtime version 5.2 is no longer available for selection in the selectionlist when creating a new cluster. 5.3 is the current oldest 5.X version you can select. There are 4 different 5.3 runtimes available, 5.3 with or without GPU and 5.3 ML with or without GPU. Please specify which one to choose in the context of the article and perhaps a link to documentation explaining the differences.