Use proper Application ID instead of Object ID
The current example for `Give Key Vault access to your storage account` is not working as intended. The Object ID for the public cloud Azure Key Vault is wrong. In addition, the approach starts by listing the Application ID instead of the Object ID, which is misleading.

With my proposed change, the example uses the Application ID and the correct command line option.