Update quickstart-v2-javascript.md
Unless I include /index.html in the "Redirect URI" section, when running the IIS/Visual Studio version of the app I obtain the error below: 
invalid_request: The provided value for the input parameter 'redirect_uri' is not valid. The expected value is a URI which matches a redirect URI registered for this client application.