Update SLES HA with socat configuration
As per latest recommendation from SUSE (https://www.suse.com/support/kb/doc/?id=7024128), have edited steps to replace "nc" resource configuration with "socat" resource. Adding socat installation steps in https://docs.microsoft.com/en-us/azure/virtual-machines/workloads/sap/high-availability-guide-suse-pacemaker and socat resource configuration in this document.

This recommendation is related to replacing “nc” resource in the SUSE HA cluster. nc resource is used for listening to the Azure LoadBalancer. Without this functionality the Floating Cluster IP is not reachable. But in some of the scenarios, it was revealed that due to the backlog in nc being hardcoded to 1 and the general limitations of the nc, this can lead to the resource being blocked. To make sure we do not run into this situation, SUSE has suggested to use socat which is more powerful and reliable than nc.