Clearer distinction but public/confidential flow
Redeeming the auth code for a token and refreshing an access token differ when you are on a confidential or public client scenario. This update tries to make difference clearer in the use of the client secret on those calls.

Disclaimer: this is merely a docs issue, I've tested both public and confidential flows and the service works as expected.