Most policy operators only support strings
Try to make it more clear that most policy operators only support strings.
Feel free to show this in a different way.

Rationale: I've seen a couple of cases recently where customers were trying to compare objects with "equals". This is not supported: the underlying function is a string comparison, and it only converts value types (e.g. bools and ints) to strings on the fly. Policy definition validation error messages could be better here, but hoping to stop a few people from trying to author such policies in the first place.