Confusing table for GRID drivers
It's difficult to see that GRID drivers should be used for Windows 10 as well. Two drivers versions appears next to Windows server 2016 and Windows server 2019, but nothing next to Windows 10. Perhaps make the drivers centered in the right column so it appears to apply to all operating systems and not just 2016/2019 server.