Updated code snippet to use `IAsyncCollector`
As part of this PR, I have replaced `ICollector` has been with `IAsyncCollector`

**Why I feel this change is important**

Whether we like it or not, most of the developers (like me) copy the code snippet directly and then modify it as per scenario. Hence, I feel it is important for docs to reflect the best possible in option in the code sample. 
I'm on the opinion that providing an `async` option over synchronous should be preferred in the snippet. While there is a mention of `IAsyncCollector` in the documentation, no written docs can take can replace the code snippet. This has motivated me to propose this change.