Include App Secret on Get Token and Refresh Token
B2C Auth code flow can be used for both confidential and public clients. In public clients, the App does not pass a client secret to refresh the token (nor to redeem auth code), whereas in confidential client it does.

Disclaimer: this is merely a docs issue, I've tested both public and confidential flows and the service works as expected.