Update control-flow-webhook-activity.md
I have updated the page with information on how to use the Report status on callback feature with information provided by Monica F on the engineering team. I have also added a little tip about the content type as it is not clear when this fails what the reason is unless you go look at what http error 415 means. Even that only means invalid content type. Lastly I removed the piece about callback on the body of the request as that body refers to the body of any endpoint you are calling not the body of the callback