Added a section on server decommissioning rather than entire product decomissioning
These steps cover the bits needed to remove a single server (decommissioning) and are based on the later section on entire removal of the product. In the server decom scenario though there is less stuff to remove. You need to uninstall the software and remove one SCP and not the forest config.