Updated the ssh user name
Updating the username for which the ssh key is getting created/applied. I had a couple of instances where resetting the azureuser's ssh key resulted in AKS cluster losing connection with Azure ARM and the nodes getting shown as unhealthy. 
So for the SSH access purpose, it's better to use a new admin account name, so that the existing connections will not be distrubed and a new user account with sufficient rights will get created for your use.