Update the exclusion for trusted location
MS Support Ticket 119122324002663 provided us with additional information for excluding locations from our sign-in risk policy.  We were advised that if we created a location with the IP flagged at risk for Unfamiliar Locations and marked the location as trusted, the risk policy would no longer trigger on the trusted location.