Added a Windows OS (Non-server) section
Filing on behalf of Japanese Internal Customer:
Context: ASC installed the MMAgent on a Win 10 VM on a end customer's subscription which they applied ASC Standard Tier to. The end customer did not see any mention of Win 10 being a supported platform and so did not think to opt-out this VM.  I made a suggested edit in this doc by including information from the “Windows” sub heading here:
https://docs.microsoft.com/en-us/azure/security-center/security-center-alerts-iaas#windows-

under a new sub heading:  ### Windows operating systems <a name="os-windows (non-server)"></a> under # Supported platforms 

But I believe much can be done to improve on this.  Any expert assistance to incorporate this above info in a way so that "customers aren't surprised to see new unexpected costs generated because ASC also installs the MMAgent on all the Non-Server Windows based VMs if auto provisioning is enabled", is greatly appreciated.